Refactor graphics package
