+++ graphics/board.py
@@ -1,8 +1,7 @@
 import pygame as pg
 from constants import Spaces
+from graphics.tile import Tile
 
-LEFT_CLICK = 1
-RIGHT_CLICK = 3
 
 class Board:
 
@@ -33,7 +32,6 @@ class Board:
         self.tiles = {}
         dark = False
 
-        # TODO: Add a different method of assigning bombs to tiles?
         for x in range(0, self.screen_width, tile_dim):
             for y in range(0, self.screen_height, tile_dim):
                 color = GREY if dark else WHITE
@@ -51,6 +49,11 @@ class Board:
         return (x // self.tile_dim, y // self.tile_dim)
 
     def update_board(self, board_array):
+        """
+        Update the board using the 2D display array from the engine
+        :param board_array:
+        :return:
+        """
         y_dim, x_dim = board_array.shape
 
         for x in range(x_dim):
@@ -68,57 +71,3 @@ class Board:
             tile.draw(self.screen)
 
         pg.display.flip()
-
-
-class Tile:
-    """
-    Represents a square on the minesweeper grid
-    """
-
-    def __init__(self, x, y, color, width, font, value):
-        """
-        Tile constructor
-        :param x: x coordinate on screen
-        :param y: y coordinate on screen
-        :param color: Default tile color
-        :param width: tile width and height
-        :param font: Font object; only passed to the object for optimization
-        """
-
-        self.width = width
-        self.x = x
-        self.y = y
-        self.rect = pg.rect.Rect(self.x, self.y, self.width, self.width)
-        self.color = (155, 155, 155)
-        self.stroke = 1
-        self.DEFAULT_COLOR = color
-        self.active = True
-        self.font = font
-        self.text_surface = self.font.render("", False, (255, 255, 255))
-        self.value = value
-
-    def set_value(self, val_enum):
-        self.value = val_enum
-        self._update_text_surface()
-
-        if val_enum != Spaces.UNKNOWN:
-            self.stroke = 0
-        else:
-            self.stroke = 1
-
-    def draw(self, screen):
-        """
-        Draw the tile in screen
-        :param screen:
-        :return:
-        """
-        pg.draw.rect(screen, self.color, self.rect, self.stroke)
-        screen.blit(self.text_surface, (self.x, self.y))
-
-    def _update_text_surface(self):
-        """
-        Update the text rendered on the tile
-        :param tile_enum:
-        :return:
-        """
-        self.text_surface = self.font.render(str(self.value), False, (255, 255, 255))
